Move handlers to `utils/exception_handlers.py`
